Extract Menu types to their own file
