+++ src/types/menu.ts
@@ -1,0 +1,9 @@
+export interface MenuItem {
+  description: string;
+}
+
+export interface Menu {
+  mains: MenuItem[];
+
+  desserts: MenuItem[];
+}
